Added an extremely hacky solution for hiding the Android action bar icons
I only want to show the icons on iOS but thus far I have not found a
way to do that. If I don’t include them in the Android resources the
app crashes with a resource not found error. I’m tired of messing with
it so to get around it I added a blank.png with matching names for the
icons in the Android resources. If anyone has a better solution I am
all ears.